Fixing some issues in ams and fms
1. Change in balance leaves to allow as fraction
2. Was assigning barcode(number) to user object reference for early checkout cases. Corrected it.
3. Sort pending leave request by date on ams_display page
4. Display total leaves taken on leave status page
5. Display scholarship info on fee, fee_report page
6. Highlight defaulters using bold font instead of red color
7. On ams,ams_display pages vertically align all lists to top

+++ ams/models.py
@@ -163,7 +163,7 @@ class Leaves(models.Model):
 class LeavesBalance(models.Model):
 	Barcode = models.ForeignKey(User)
 	Type = models.PositiveIntegerField(choices=LEAVE_CHOICES)
-	Days = models.PositiveIntegerField()
+	Days = models.FloatField()
 
 class LeaveRules(models.Model):
 	Category = models.ForeignKey(Category)
